Added second approach for dealing with web3 provider network aware
